[MOVE] GetProfileAwardDto 위치 변경 및 Swagger 추가
