+++ src/main/java/synk/meeteam/domain/user/user/api/UserController.java
@@ -15,9 +15,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import synk.meeteam.domain.common.skill.dto.SkillDto;
 import synk.meeteam.domain.portfolio.portfolio.dto.GetProfilePortfolioDto;
+import synk.meeteam.domain.user.award.dto.GetProfileAwardDto;
 import synk.meeteam.domain.user.user.dto.request.UpdateProfileRequestDto;
 import synk.meeteam.domain.user.user.dto.response.CheckDuplicateNicknameResponseDto;
-import synk.meeteam.domain.user.user.dto.response.GetProfileAwardDto;
 import synk.meeteam.domain.user.user.dto.response.GetProfileEmailDto;
 import synk.meeteam.domain.user.user.dto.response.GetProfileLinkDto;
 import synk.meeteam.domain.user.user.dto.response.GetProfilePhoneDto;
