--- src/main/java/synk/meeteam/domain/user/user/dto/response/GetProfileAwardDto.java
@@ -1,13 +1,0 @@
-package synk.meeteam.domain.user.user.dto.response;
-
-import java.time.LocalDate;
-
-public record GetProfileAwardDto(
-        String title,
-        LocalDate startDate,
-        LocalDate endDate,
-        String description
-) {
-
-
-}
